feat: rust 1.65 -> 1.66

+++ examples/multi-machine-libp2p.rs
@@ -404,7 +404,7 @@ impl Config {
         let initializer = hotshot::HotShotInitializer::from_genesis(genesis_block).unwrap();
 
         // Create the initial hotshot
-        let known_nodes: Vec<_> = (0..self.num_nodes as u64)
+        let known_nodes: Vec<_> = (0..self.num_nodes)
             .map(|x| {
                 let priv_key = BlsPubKey::generate_test_key(x);
                 BlsPubKey::from_private(&priv_key)
@@ -430,7 +430,7 @@ impl Config {
         let hotshot = HotShot::init(
             self.pubkey.clone(),
             self.privkey.clone(),
-            self.node_id as u64,
+            self.node_id,
             config,
             networking,
             MemoryStorage::new(),
